+++ src/el/sayid-pkg.el
@@ -1,4 +1,0 @@
-(define-package
-    "sayid"
-    "0.0.11"
-  "An nrepl middleware client for sayid, a clojure debugger.")