+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Change Log
 
-## [0.0.11] - 2016-01-10
+## [0.0.14] - 2017-03-06
+### Changed
+ - emacs: several misc things for MELPA
+
+## [0.0.13] - 2017-02-22
+### Fixed
+- emacs: messed up some function names
+
+## [0.0.12] - 2017-02-21
+### Changed
+- emacs: improve render speed
+- emacs: prepare package for submission to melpa
+- emacs: useful message when sayid not responding
+
+### Fixed
+- support sets in pretty-print buffer
+
+## [0.0.11] - 2017-01-10
 ### Added
 - CHANGELOG.md